requirements/base.txt
tzdata==2022.1
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==6.2.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.4  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
django-celerybeat-status==0.0.10 # https://github.com/vintasoftware/django-celerybeat-status
django-celery-results==1.1.2 # https://pypi.org/project/django-celery-results/
flower==0.9.3  # https://github.com/mher/flower
vine==1.3.0

# Django
# ------------------------------------------------------------------------------
django==2.2.7  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.40.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-nested-admin==3.2.4

# Oauth2
django-oauth-toolkit==1.2.0
django-cors-middleware==1.5.0
djangorestframework-simplejwt==4.3.0
pyjwt==1.7.1

django-jet==1.0.8
django_jsoneditor==0.1.4
